전체 도시락 목록 조회

+++ src/main/java/likelion/senifood/repository/LunchboxRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LunchboxRepository extends JpaRepository<Lunchbox, Integer> {
+public interface LunchboxRepository extends JpaRepository<Lunchbox, Long> {
 }
